replaced rewards widget with discord, create /dai backup link
