+++ app/earlyAccessSuccess/page.tsx
@@ -3,15 +3,15 @@
 import Image from "next/image";
 import React, { useEffect } from "react";
 
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { useRouter, useSearchParams } from "next/navigation";
 
-import TrophyIcon from "../../public/TrophyIcon.svg";
-import CalendarIcon from "../../public/CalendarIcon.svg";
-import StayConnectedIcon from "../../public/StayConnectedIcon.svg";
+import TrophyIcon from "@/public/TrophyIcon.svg";
+import CalendarIcon from "@/public/CalendarIcon.svg";
+import StayConnectedIcon from "@/public/StayConnectedIcon.svg";
 
-import MissionImageNew from "../../public/MissionImageNew.webp";
-import CalendarImage from "../../public/CalendarImage.webp";
-import ConnectedWorldImage from "../../public/ConnectedWorldImage.webp";
+import MissionImageNew from "@/public/MissionImageNew.webp";
+import CalendarImage from "@/public/CalendarImage.webp";
+import ConnectedWorldImage from "@/public/ConnectedWorldImage.webp";
 
 import {
   FaDiscord,
